refactor(Server): delete createFeatures.js
